# agentbox-client/openapi/agentbox-manager.openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "sandbox_id": {
            "title": "Sandbox Id",
            "type": "string"
          }
        },
        "required": [
          "sandbox_id",
          "deleted"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SandboxEnsureRequest": {
        "additionalProperties": false,
        "properties": {
          "env": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Env",
            "type": "object"
          }
        },
        "title": "SandboxEnsureRequest",
        "type": "object"
      },
      "SandboxResponse": {
        "properties": {
          "sandbox": {
            "$ref": "#/components/schemas/SandboxSummary"
          }
        },
        "required": [
          "sandbox"
        ],
        "title": "SandboxResponse",
        "type": "object"
      },
      "SandboxSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "CREATING",
              "RUNNING",
              "STOPPED",
              "ERROR"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ready",
          "status"
        ],
        "title": "SandboxSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentBox Manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sandboxes/{sandbox_id}": {
      "delete": {
        "operationId": "delete_sandboxes_sandbox_id",
        "parameters": [
          {
            "in": "path",
            "name": "sandbox_id",
            "required": true,
            "schema": {
              "title": "Sandbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Sandbox"
      },
      "get": {
        "operationId": "get_sandboxes_sandbox_id",
        "parameters": [
          {
            "in": "path",
            "name": "sandbox_id",
            "required": true,
            "schema": {
              "title": "Sandbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sandbox"
      },
      "put": {
        "operationId": "put_sandboxes_sandbox_id",
        "parameters": [
          {
            "in": "path",
            "name": "sandbox_id",
            "required": true,
            "schema": {
              "title": "Sandbox Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxEnsureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ensure Sandbox"
      }
    },
...
  }
}